Fixing some stuff with the mini tutorial
